Combined thumbnail / browse queueing into derived queueing
