+++ recipe/new/mro_ctx.json
@@ -84,23 +84,32 @@
                 "from_": "{{no_extension_inputfile}}.cub",
                 "to": "{{no_extension_inputfile}}.cal.cub"
             },
-            "gdal_translate": {
+            "gdal.gdal_translate": {
             		"outputType": "Byte",
             		"format" : "JPEG",
-                        "scaleParams": [[]],
-            		"width" : "{{width}}" ,
-            		"height" : "{{height}}",
+            		"scaleParams": [[]],
+            		"width" : "0",
+            		"height" : "300",
             		"src" : "{{no_extension_inputfile}}.cal.cub",
-            		"dest": "{{derived_product}}"
-      	    }
+            		"dest": "{{derived_product}}.thumbnail.jpg"
+      	    },
+            "gdal.gdal_translate": {
+                "outputType": "Byte",
+                "format" : "JPEG",
+                "scaleParams": "",
+                "width" : "0",
+                "height" : "900",
+                "src" : "{{no_extension_inputfile}}.cal.cub",
+                "dest": "{{derived_product}}.browse.jpg"
+            }
         },
 	"browse": {
 	    "width": "0",
-	    "height": "900"
+	    "height": "900",
 	},
 	"thumbnail": {
-	    "width": "0",
-	    "height": "300"
+	    "width": "0"
+	    "height": "300",
 	}
     },
     "projected": {
